Use `@Option` instead of inheritance
